채팅기능 중간저장
