+++ src/App.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: "KOTRAHOPE";
-  src: url("https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2110@1.0/KOTRAHOPE.woff2")
+  src: url("https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts/211001_kotrahope.woff2")
     format("woff2");
   font-weight: normal;
   font-style: normal;
