+++ src/components/Chat/MyChatDummy.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const MyChatDummy = [
+  { id: 1, content: "ㅎㅇ", date: "2022-01-01" },
+  { id: 2, content: "ㅋㅋ", date: "2022-01-02" },
+  { id: 3, content: "ㄱㄱ", date: "2022-01-01" },
+  { id: 4, content: "ㄴㄴ", date: "2022-01-02" },
+];
+
+export default MyChatDummy;
