--- src/components/MyChatDummy.tsx
@@ -1,31 +1,0 @@
-import React from "react";
-
-const MyChatDummy = [
-  { id: 1, content: "ㅎㅇ", date: "2022-01-01" },
-  { id: 2, content: "ㅋㅋ", date: "2022-01-02" },
-  { id: 3, content: "ㄱㄱ", date: "2022-01-01" },
-  { id: 4, content: "ㄴㄴ", date: "2022-01-02" },
-];
-
-export default MyChatDummy;
-
-{
-  /* <button
-onClick={() => {
-  axios
-    .get("https://codingapple1.github.io/shop/data2.json")
-    .then((res) => {
-      console.log(res.data);
-
-      let copy = [...shoes, ...res.data];
-
-      setShoes(copy);
-    })
-    .catch(() => {
-      console.log("에러뜸");
-    });
-}}
->
-그버튼
-</button> */
-}
